requirements.txt
torch>=2.2.0
torchvision>=0.17.0
pillow>=10.0.0
pandas>=2.0.0
scikit-learn>=1.3.0
streamlit>=1.37.0

